Fixed VU rendering bug in the PS2 (error didn't show on the emulator)
Found out what the drawing issue was on the PS2. Was missing a
synchronization after each batch to prevent writing to a VU memory
location that was still in use. Fix was to just send a wait finish
event to the GS and wait for draw to complete before sending a new
batch to the VU1.